Fluid/requirements.txt
docker
python-etcd
Flask==2.2.2
Flask_RESTful==0.3.10
inotify==0.2.10
inotify_simple==1.3.5
Requests==2.32.3
